Fix Null Reference warning
